backend/requirements.txt
# Redstone Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Redis
redis>=5.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# HTTP Client
httpx>=0.26.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# LLM APIs
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.4.0

# Stripe
stripe>=7.10.0

# Utilities
python-dotenv>=1.0.0
